requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.10.1
# MySQL 驅動 (開發環境)
PyMySQL==1.1.2
# PostgreSQL 驅動 (生產環境)
psycopg2-binary==2.9.9
